Implemented distinct Likes functionality. Refactored a couple of dtos and repositories.

+++ src/main/java/com/example/musicapi/dtos/song_dtos/SongDto.java
@@ -9,7 +9,7 @@ public record SongDto(
         String title,
         String author,
         Date addedAt,
-        Long likes,
+        int likes,
         byte[] data,
         Long duration,
         MusicGenre genre) {
